Функциональность «Поиск». Исправления по замечаниям.

+++ src/main/java/ru/yandex/practicum/filmorate/model/Director.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Builder
 public class Director {
 
-    private final Integer id;
+    private final long id;
 
     private final String name;
 }
